requirements.txt
opencv-python
tqdm
tensorboard
addict
scikit-learn
pathspec
numpy==1.19.4; python_version<"3.7"
numpy==1.20.2; python_version>="3.7"
# numpy==1.23.1   

# dev-dep
yapf 

# deploy
# protobuf==3.20.0
onnx==1.14.0
onnxruntime==1.9.0
onnx-simplifier

